+++ src/components/Grid/SimpleGrid.test.tsx
@@ -2,7 +2,6 @@ import { render } from "@testing-library/react";
 import { axe } from "jest-axe";
 import * as React from "react";
 import renderer from "react-test-renderer";
-import { mockAllIsIntersecting } from "react-intersection-observer/test-utils";
 
 import Card, { CardHeading, CardContent } from "../Card/Card";
 import SimpleGrid from "./SimpleGrid";
@@ -52,8 +51,6 @@ describe("Grid Accessibility", () => {
         </Card>
       </SimpleGrid>
     );
-    // Mock IntersectionObserver to render images.
-    mockAllIsIntersecting(true);
 
     expect(await axe(container)).toHaveNoViolations();
   });