+++ src/components/StructuredContent/StructuredContent.test.tsx
@@ -2,7 +2,6 @@ import { render, screen } from "@testing-library/react";
 import { axe } from "jest-axe";
 import * as React from "react";
 import renderer from "react-test-renderer";
-import { mockAllIsIntersecting } from "react-intersection-observer/test-utils";
 
 import StructuredContent from "./StructuredContent";
 
@@ -79,11 +78,6 @@ const htmlDOMBodyContent = (
 );
 
 describe("StructuredContent Accessibility", () => {
-  beforeEach(() => {
-    // Mock IntersectionObserver to render images.
-    mockAllIsIntersecting(true);
-  });
-
   it("passes axe accessibility for HTML string text content", async () => {
     const { container } = render(
       <StructuredContent
@@ -156,8 +150,6 @@ describe("StructuredContent", () => {
     );
     const mainHeading = screen.getByRole("heading", { level: 2 });
     const calloutHeading = screen.getByRole("heading", { level: 3 });
-    // Mock IntersectionObserver to render images.
-    mockAllIsIntersecting(true);
 
     expect(mainHeading).toHaveTextContent("Heading text");
     expect(calloutHeading).toHaveTextContent("This is the callout text");
@@ -182,8 +174,6 @@ describe("StructuredContent", () => {
     );
     const mainHeading = screen.queryByRole("heading", { level: 2 });
     const calloutHeading = screen.queryByRole("heading", { level: 3 });
-    // Mock IntersectionObserver to render images.
-    mockAllIsIntersecting(true);
 
     expect(mainHeading).not.toBeInTheDocument();
     expect(calloutHeading).not.toBeInTheDocument();
@@ -225,8 +215,6 @@ describe("StructuredContent", () => {
         }}
       />
     );
-    // Mock IntersectionObserver to render images.
-    mockAllIsIntersecting(true);
 
     expect(warn).toHaveBeenCalledWith(
       "NYPL Reservoir StructuredContent: The `imageProps.alt` prop is required " +
@@ -252,8 +240,6 @@ describe("StructuredContent", () => {
         }}
       />
     );
-    // Mock IntersectionObserver to render images.
-    mockAllIsIntersecting(true);
 
     expect(screen.getByRole("img")).toBeInTheDocument();
     expect(screen.getByRole("figure")).toBeInTheDocument();
@@ -297,8 +283,6 @@ describe("StructuredContent", () => {
         }}
       />
     );
-    // Mock IntersectionObserver to render images.
-    mockAllIsIntersecting(true);
 
     expect(screen.getByRole("img")).toBeInTheDocument();
     expect(screen.queryByRole("figure")).not.toBeInTheDocument();
@@ -440,8 +424,6 @@ describe("StructuredContent", () => {
         />
       )
       .toJSON();
-    // Mock IntersectionObserver to render images.
-    mockAllIsIntersecting(true);
 
     expect(withHTMLStringContent).toMatchSnapshot();
     expect(withHTMLDOMContent).toMatchSnapshot();